fix(material): D-466 品质异常数量与库存告警数量改小数（物料侧收官）
按确认口径：成品衣服按件保持整数，物料（面料/辅料/里料）一律小数。

1) 品质异常数量（涉及库存扣减，属于真缺口）
- 迁移 V202709181900：t_material_quality_issue.issue_quantity INT → DECIMAL(12,4)
- 实体 MaterialQualityIssue.issueQuantity → BigDecimal
- 新建异常：intOf → decimalOf（原会把 1.5 米截成 1），上限比较改 BigDecimal
- 异常回退到货量、生成补料采购单、异常金额三处去掉 BigDecimal.valueOf(整数) 包装，
  直接按小数参与计算（此前异常 1.5 米只回退/补料/计钱 1 米）
- 前端异常数量输入：precision={0} → precision={2} step={0.01}，min 1 → 0.01

2) 库存告警数量（展示层）
- MaterialStockAlertDto.quantity → BigDecimal，去掉 intValue() 取整
  （375.5 米此前显示成 375）
- shortage() 缺口 = 建议安全库存 - 当前库存，改按小数计算后向上取整
  （缺 0.5 米应提示补 1，不能提示补 0）

自检：mvn compile / test-compile 通过、tsc 0 错、eslint 0 error、npm run build 通过；
未触碰打印相关文件。

+++ backend/src/main/java/com/fashion/supplychain/production/dto/MaterialStockAlertDto.java
@@ -14,7 +14,8 @@ public class MaterialStockAlertDto {
     private String unit;
     private String color;
     private String size;
-    private Integer quantity;
+    /** 当前库存（D-466：改为小数，面料 375.5 米不应显示成 375） */
+    private BigDecimal quantity;
     private Integer safetyStock;
     private Integer recentOutQuantity;
     private Integer suggestedSafetyStock;
